Refactory: 코드 리뷰 반영
- 응답_조회 메소드를 별도의 스텝 메소드를 가진 클래스로 추출
- 스텝 메소드의 의미를 명확하도록 수정
- 조회 트랜잭션에는 readOnly 추가
